--- a/src/lib.rs
+++ b/src/lib.rs
@@ -41,7 +41,7 @@ pub(crate) fn compress<C: AffineRepr>(vec: &mut Vec<C>, split: usize, scaler: &C
         .zip(right.par_iter())
         .for_each(|(a_l, a_r)| {
             //let mut x = mul!(a_r.into_group(), scaler.clone());
-            let sc = scaler.clone();
+            let sc = *scaler;
             let mut x = a_r.mul(sc);
             x.add_assign(*a_l);
             *a_l = x.into_affine();