- fixes most of the clippy warnings
- disables some warnings about complex types or too many arguments in par with the bellperson crate
- also disables the wrongly linted clones in the verifyer to which issue https://github.com/nikkolasg/snarkpack/issues/6 is related.
- configure the CI to deny clippy warnings

Let's test the benchmarking bot :)