@nikkolasg 2023-06-11
## Benchmark for 143c6f2
  <details>
    <summary>Click to view benchmark</summary>

| Test | Base         | PR               | % |
|------|--------------|------------------|---|
| groth16 aggregation | 1463.7±25.24ms | 1462.9±26.65ms | -0.05% |

  </details>