@nikkolasg 2023-06-11
## Benchmark for 143c6f2
  <details>
    <summary>Click to view benchmark</summary>

| Test | Base         | PR               | % |
|------|--------------|------------------|---|
| groth16 aggregation | 1609.8±17.65ms | 1618.1±24.62ms | +0.52% |

  </details>