@achimcc 2023-06-19
Need to think a bit more if it is justified to remove the references from the concurrent parts...